packages/jobs/dev-requirements.txt
-r requirements.txt

pytest==6.1.2
pytest-mock==3.5.1
pytest-json==0.4.0
flake8==3.8.4
black==20.8b1
parsel==1.6.0
python-dateutil==2.8.2
